Move malloc functions into separate file
