+++ src/slap/matrix.c
@@ -1,36 +1,13 @@
 #include "matrix.h"
 
 #include <math.h>
-#include <stddef.h>
-#include <stdlib.h>
-#include <string.h>
+#include <memory.h>
 
 #include "matrix_checks.h"
 
 Matrix slap_MatrixFromArray(int rows, int cols, double* data) {
   Matrix mat = {rows, cols, 1, rows, data, slap_DENSE};
   return mat;
-}
-
-Matrix slap_NewMatrix(int rows, int cols) {
-  double* data = (double*)malloc(rows * cols * sizeof(double));
-  Matrix mat = {rows, cols, 1, rows, data, slap_DENSE};
-  return mat;
-}
-
-Matrix slap_NewMatrixZeros(int rows, int cols) {
-  double* data = (double*)calloc(rows * cols, sizeof(double));
-  Matrix mat = {rows, cols, 1, rows, data, slap_DENSE};
-  return mat;
-}
-
-int slap_FreeMatrix(Matrix mat) {
-  if (mat.data) {
-    free(mat.data);
-    mat.data = NULL;
-    return SLAP_NO_ERROR;
-  }
-  return SLAP_BAD_MATRIX_DATA_POINTER;
 }
 
 enum slap_ErrorCode slap_SetConst(Matrix mat, double val) {
